feat(cmake): add option to conditionally build C lcmtypes
The LCMTYPES_BUILD_C option is added to the CMake configuration to
conditionally build C lcmtypes. This option is set to ON by default,
allowing the C lcmtypes to be built unless explicitly disabled by the
user. This change provides flexibility for users who may not require
the C lcmtypes in their projects, allowing them to save build time and
have a less polluted build environment (e.g. the include folder).